renamed pus_tc to tc and pus_tm to tm
