+++ docs/api/tmtccmd.pus_tc.rst
@@ -1,69 +1,69 @@
-tmtccmd.pus\_tc package
+tmtccmd.tc package
 =======================
 
 Submodules
 ----------
 
-tmtccmd.pus\_tc.definitions module
+tmtccmd.tc.definitions module
 ----------------------------------
 
-.. automodule:: tmtccmd.pus_tc.definitions
+.. automodule:: tmtccmd.tc.definitions
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tc.packer module
+tmtccmd.tc.packer module
 -----------------------------
 
-.. automodule:: tmtccmd.pus_tc.packer
+.. automodule:: tmtccmd.tc.packer
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tc.service\_17\_test module
+tmtccmd.tc.service\_17\_test module
 ----------------------------------------
 
-.. automodule:: tmtccmd.pus_tc.service_17_test
+.. automodule:: tmtccmd.tc.service_17_test
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tc.service\_200\_mode module
+tmtccmd.tc.service\_200\_mode module
 -----------------------------------------
 
-.. automodule:: tmtccmd.pus_tc.service_200_mode
+.. automodule:: tmtccmd.tc.service_200_mode
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tc.service\_20\_parameter module
+tmtccmd.tc.service\_20\_parameter module
 ---------------------------------------------
 
-.. automodule:: tmtccmd.pus_tc.service_20_parameter
+.. automodule:: tmtccmd.tc.service_20_parameter
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tc.service\_3\_housekeeping module
+tmtccmd.tc.service\_3\_housekeeping module
 -----------------------------------------------
 
-.. automodule:: tmtccmd.pus_tc.service_3_housekeeping
+.. automodule:: tmtccmd.tc.service_3_housekeeping
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tc.service\_5\_event module
+tmtccmd.tc.service\_5\_event module
 ----------------------------------------
 
-.. automodule:: tmtccmd.pus_tc.service_5_event
+.. automodule:: tmtccmd.tc.service_5_event
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tc.service\_8\_functional\_cmd module
+tmtccmd.tc.service\_8\_functional\_cmd module
 --------------------------------------------------
 
-.. automodule:: tmtccmd.pus_tc.service_8_functional_cmd
+.. automodule:: tmtccmd.tc.service_8_functional_cmd
    :members:
    :undoc-members:
    :show-inheritance:
@@ -71,7 +71,7 @@ tmtccmd.pus\_tc.service\_8\_functional\_cmd module
 Module contents
 ---------------
 
-.. automodule:: tmtccmd.pus_tc
+.. automodule:: tmtccmd.tc
    :members:
    :undoc-members:
    :show-inheritance:
