+++ docs/api/tmtccmd.pus_tm.rst
@@ -1,93 +1,93 @@
-tmtccmd.pus\_tm package
+tmtccmd.tm package
 =======================
 
 Submodules
 ----------
 
-tmtccmd.pus\_tm.definitions module
+tmtccmd.tm.definitions module
 ----------------------------------
 
-.. automodule:: tmtccmd.pus_tm.definitions
+.. automodule:: tmtccmd.tm.definitions
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.factory module
+tmtccmd.tm.factory module
 ------------------------------
 
-.. automodule:: tmtccmd.pus_tm.factory
+.. automodule:: tmtccmd.tm.factory
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_17\_test module
+tmtccmd.tm.service\_17\_test module
 ----------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_17_test
+.. automodule:: tmtccmd.tm.service_17_test
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_1\_verification module
+tmtccmd.tm.service\_1\_verification module
 -----------------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_1_verification
+.. automodule:: tmtccmd.tm.service_1_verification
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_200\_mode module
+tmtccmd.tm.service\_200\_mode module
 -----------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_200_mode
+.. automodule:: tmtccmd.tm.service_200_mode
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_20\_parameters module
+tmtccmd.tm.service\_20\_parameters module
 ----------------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_20_parameters
+.. automodule:: tmtccmd.tm.service_20_parameters
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_2\_raw\_cmd module
+tmtccmd.tm.service\_2\_raw\_cmd module
 -------------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_2_raw_cmd
+.. automodule:: tmtccmd.tm.service_2_raw_cmd
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_3\_base module
+tmtccmd.tm.service\_3\_base module
 ---------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_3_base
+.. automodule:: tmtccmd.tm.service_3_base
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_3\_housekeeping module
+tmtccmd.tm.service\_3\_housekeeping module
 -----------------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_3_housekeeping
+.. automodule:: tmtccmd.tm.service_3_housekeeping
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_5\_event module
+tmtccmd.tm.service\_5\_event module
 ----------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_5_event
+.. automodule:: tmtccmd.tm.service_5_event
    :members:
    :undoc-members:
    :show-inheritance:
 
-tmtccmd.pus\_tm.service\_8\_functional\_cmd module
+tmtccmd.tm.service\_8\_functional\_cmd module
 --------------------------------------------------
 
-.. automodule:: tmtccmd.pus_tm.service_8_functional_cmd
+.. automodule:: tmtccmd.tm.service_8_functional_cmd
    :members:
    :undoc-members:
    :show-inheritance:
@@ -95,7 +95,7 @@ tmtccmd.pus\_tm.service\_8\_functional\_cmd module
 Module contents
 ---------------
 
-.. automodule:: tmtccmd.pus_tm
+.. automodule:: tmtccmd.tm
    :members:
    :undoc-members:
    :show-inheritance:
