+++ example/config/hook_implementation.py
@@ -12,9 +12,9 @@ class ExampleHookClass(TmTcHookBase):
     from tmtccmd.core.backend import TmTcHandler
     from tmtccmd.utility.tmtc_printer import TmTcPrinter
     from tmtccmd.ecss.tm import PusTelemetry
-    from tmtccmd.pus_tc.definitions import TcQueueT
+    from tmtccmd.tc.definitions import TcQueueT
     from tmtccmd.com_if.com_interface_base import CommunicationInterface
-    from tmtccmd.pus_tm.service_3_base import Service3Base
+    from tmtccmd.tm.service_3_base import Service3Base
 
     def get_json_config_file_path(self) -> str:
         return "tmtc_config.json"
@@ -43,14 +43,14 @@ class ExampleHookClass(TmTcHookBase):
         pass
 
     def pack_service_queue(self, service: Union[str, int], op_code: str, service_queue: TcQueueT):
-        from tmtccmd.pus_tc.packer import default_service_queue_preparation
+        from tmtccmd.tc.packer import default_service_queue_preparation
         LOGGER.info("Service queue packer hook was called")
         default_service_queue_preparation(
             service=service, op_code=op_code, service_queue=service_queue
         )
 
     def tm_user_factory_hook(self, raw_tm_packet: bytearray) -> PusTelemetry:
-        from tmtccmd.pus_tm.handler import default_factory_hook
+        from tmtccmd.tm.handler import default_factory_hook
         LOGGER.info("TM user factory hook was called")
         return default_factory_hook(raw_tm_packet=raw_tm_packet)
