+++ src/tmtccmd/com_if/dummy_com_if.py
@@ -7,9 +7,9 @@ from typing import cast
 
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
 from tmtccmd.ecss.tc import PusTelecommand
-from tmtccmd.pus_tm.definitions import PusTmListT
-from tmtccmd.pus_tm.service_1_verification import Service1TmPacked
-from tmtccmd.pus_tm.service_17_test import Service17TmPacked
+from tmtccmd.tm.definitions import PusTmListT
+from tmtccmd.tm.service_1_verification import Service1TmPacked
+from tmtccmd.tm.service_17_test import Service17TmPacked
 from tmtccmd.pus.service_17_test import Srv17Subservices
 from tmtccmd.utility.logger import get_logger
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
