+++ src/tmtccmd/sendreceive/sequential_sender_receiver.py
@@ -13,7 +13,7 @@ from tmtccmd.sendreceive.tm_listener import TmListener
 from tmtccmd.com_if.com_interface_base import CommunicationInterface
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.utility.logger import get_logger
-from tmtccmd.pus_tc.definitions import TcQueueT
+from tmtccmd.tc.definitions import TcQueueT
 
 LOGGER = get_logger()
