+++ src/tmtccmd/sendreceive/single_command_sender_receiver.py
@@ -16,7 +16,7 @@ from tmtccmd.com_if.com_interface_base import CommunicationInterface
 from tmtccmd.utility.tmtc_printer import TmTcPrinter
 from tmtccmd.utility.logger import get_logger
 
-from tmtccmd.pus_tc.definitions import PusTcTupleT
+from tmtccmd.tc.definitions import PusTcTupleT
 
 
 logger = get_logger()
