+++ src/tmtccmd/tc/definitions.py
