+++ src/tmtccmd/tc/service_5_event.py
@@ -1,6 +1,6 @@
 from tmtccmd.config.globals import get_global_apid
 from tmtccmd.config.definitions import QueueCommands
-from tmtccmd.pus_tc.definitions import PusTelecommand, TcQueueT
+from tmtccmd.tc.definitions import PusTelecommand, TcQueueT
 from tmtccmd.pus.service_5_event import Srv5Subservices
