+++ src/tmtccmd/utility/tmtc_printer.py
@@ -9,11 +9,11 @@ from typing import cast
 
 from tmtccmd.ecss.tc import PusTelecommand
 from tmtccmd.ecss.tm import PusTelemetry
-from tmtccmd.pus_tm.service_8_functional_cmd import Service8TM
-from tmtccmd.pus_tm.service_5_event import Service5TM
+from tmtccmd.tm.service_8_functional_cmd import Service8TM
+from tmtccmd.tm.service_5_event import Service5TM
 from tmtccmd.config.definitions import TelemetryQueueT
-from tmtccmd.pus_tm.definitions import PusTmListT
-from tmtccmd.pus_tm.service_3_base import Service3Base
+from tmtccmd.tm.definitions import PusTmListT
+from tmtccmd.tm.service_3_base import Service3Base
 from tmtccmd.utility.logger import get_logger
 
 LOGGER = get_logger()
